requirements.txt
Django>=3.2,<3.3
djangorestframework>=3.13,<4.0
psycopg2-binary>=2.9
redis>=4.5
django-redis>=5.2
djangorestframework-simplejwt>=5.2
python-dotenv>=1.0

